+++ CHANGELOG.md
@@ -5,6 +5,17 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 [Keep a CHANGELOG](http://keepachangelog.com).
 
 
+## Unreleased
+
+### Changed
+
+- adapt to new query result representation in SPARQL.ex 0.2
+
+
+[Compare v0.1.0...HEAD](https://github.com/marcelotto/sparql_client/compare/v0.1.0...HEAD)
+
+
+
 ## v0.1.0 - 2018-03-19
 
 Initial release